scripts/dashboard/app/requirements.txt
# ARTEK AI Worker Dashboard Requirements

# Core
streamlit>=1.52.0
python-dotenv>=1.2.0

# YAML parsing for routes
PyYAML>=6.0.3

# Data handling
pandas>=2.3.0

# API requests (for Cloudflare D1)
requests>=2.32.0

